.....
git-svn-id: svn://fileserver/activex/AVS/Sources/TeamlabOffice/trunk/ServerComponents@57848 954022d7-b5bf-4e40-9824-e11837661b57

+++ ASCOfficeOdfFileW/source/OdfFormat/odf_drawing_context.h
@@ -111,7 +111,7 @@ public:
 	void set_flip_H(bool bVal);
 	void set_flip_V(bool bVal);
 
-	void set_rotate(int iVal);
+	void set_rotate(double iVal);
 
 	void set_no_fill();
 	void set_solid_fill(std::wstring hexColor);
